refactor(sync): move tombstone and metadata steps to their phase files
engine_export.go declared an export boundary but also hosted two
functions that do not belong to it. processTombstones is the push
phase's remote-delete step, so it now lives in engine_push.go beside
push, where the push-phase reader finds it. syncCalendarMetadata is
calendar color/name reconciliation: it is neither export, push, nor
pull, so it moves to a new engine_metadata.go whose top comment
states that boundary.

Pure code motion: bodies are byte-identical, receivers and
signatures unchanged. engine_export.go drops the now-unused caldav
and storage imports; each touched file keeps exactly the imports
it uses.
